+++ src/com/jpexs/decompiler/flash/gui/tagtree/TagTreeModel.java
@@ -102,6 +102,11 @@ public class TagTreeModel extends AbstractTagTreeModel {
 
     public static final String FOLDER_SCENES = "scenes";
 
+    public static final String FOLDER_UNKNOWN = "unknown";
+
+    public static final String FOLDER_ERRORED = "errored";
+
+    
     public static final List<String> FOLDERS_ORDER = Arrays.asList(
             "header",
             "cookies",
@@ -118,6 +123,8 @@ public class TagTreeModel extends AbstractTagTreeModel {
             "frames",
             "scenes",
             "others",
+            "unknown",
+            "errored",
             "scripts"
     );
 
@@ -202,7 +209,9 @@ public class TagTreeModel extends AbstractTagTreeModel {
             List<TreeItem> movies,
             List<TreeItem> sounds,
             List<TreeItem> binaryData,
-            List<TreeItem> others
+            List<TreeItem> others,
+            List<TreeItem> unknown,
+            List<TreeItem> errored
     ) {
         for (Tag t : timelined.getTags()) {
             TreeNodeType ttype = TagTree.getTreeNodeType(t);
@@ -215,7 +224,7 @@ public class TagTreeModel extends AbstractTagTreeModel {
                     break;
                 case SPRITE:
                     sprites.add(t);
-                    walkTimelinedTagList((DefineSpriteTag) t, mappedTags, shapes, morphShapes, sprites, buttons, images, fonts, texts, movies, sounds, binaryData, others);
+                    walkTimelinedTagList((DefineSpriteTag) t, mappedTags, shapes, morphShapes, sprites, buttons, images, fonts, texts, movies, sounds, binaryData, others, unknown, errored);
                     break;
                 case BUTTON:
                     buttons.add(t);
@@ -240,6 +249,12 @@ public class TagTreeModel extends AbstractTagTreeModel {
                     break;
                 case AS:
                 case AS_FRAME:
+                    break;
+                case UNKNOWN:
+                    unknown.add(t);
+                    break;
+                case ERRORED:
+                    errored.add(t);
                     break;
                 default:
                     if (t.getId() != ShowFrameTag.ID && !ShowFrameTag.isNestedTagType(t.getId())) {
@@ -283,9 +298,11 @@ public class TagTreeModel extends AbstractTagTreeModel {
         List<TreeItem> sounds = new ArrayList<>();
         List<TreeItem> binaryData = new ArrayList<>();
         List<TreeItem> others = new ArrayList<>();
+        List<TreeItem> unknown = new ArrayList<>();
+        List<TreeItem> errored = new ArrayList<>();
         List<FolderItem> emptyFolders = new ArrayList<>();
         Map<Integer, List<TreeItem>> mappedTags = new HashMap<>();
-        walkTimelinedTagList(swf, mappedTags, shapes, morphShapes, sprites, buttons, images, fonts, texts, movies, sounds, binaryData, others);
+        walkTimelinedTagList(swf, mappedTags, shapes, morphShapes, sprites, buttons, images, fonts, texts, movies, sounds, binaryData, others, unknown, errored);
 
         Timeline timeline = swf.getTimeline();
         int frameCount = timeline.getFrameCount();
@@ -363,6 +380,12 @@ public class TagTreeModel extends AbstractTagTreeModel {
                     break;
                 case "others":
                     addFolderItem(nodeList, emptyFolders, true /*always add*/, translate("node.others"), FOLDER_OTHERS, swf, others);
+                    break;
+                case "unknown":
+                    addFolderItem(nodeList, emptyFolders, false, translate("node.unknown"), FOLDER_UNKNOWN, swf, unknown);
+                    break;
+                case "errored":
+                    addFolderItem(nodeList, emptyFolders, false, translate("node.errored"), FOLDER_ERRORED, swf, errored);
                     break;
                 case "scripts":
                     if (swf.isAS3()) {